+++ game/utils/Color.h
@@ -9,8 +9,8 @@
 class Color {
 public:
     Color();
-    Color(int r, int g, int b);
-    int r, g, b;
+    Color(float r, float g, float b);
+    float r, g, b;
 };
